+++ R/utils.R
@@ -1,4 +1,0 @@
-
-paste_line <- function(...) {
-  paste(c(...), collapse = "\n")
-}